pyproject.toml
[project]
name = "spamo-pjm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "av>=16.0.1",
    "datasets>=4.4.1",
    "lmdb>=1.7.5",
    "msgpack>=1.1.2",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
